show admin info by default. #362
