--- app/controllers/settings/index.js
@@ -1,4 +1,0 @@
-import Controller from '@ember/controller';
-
-export default Controller.extend({
-});
