+++ app/router.js
@@ -67,11 +67,10 @@ Router.map(function() {
       this.route('dois');
     });
   });
-  this.route('settings', function() {
-    this.route('index', { path: '/' });
-    this.route('edit');
-    this.route('change');
-  });
+  this.route('edit');
+  this.route('change');
+  this.route('dashboard');
+
   this.route('sign-in');
   this.route('authorize');
   this.route('reset');
