+++ app/routes/index.js
@@ -3,39 +3,21 @@ import { inject as service } from '@ember/service';
 
 export default Route.extend({
   can: service(),
-  currentUser: service(),
-  flashMessages: service(),
 
   model() {
-    if (this.can.can('read index')) {
-      let self = this;
-      return this.store.findRecord('provider', 'admin').then(function(admin) {
-        return admin;
-      }).catch(function(reason) {
-        console.debug(reason);
+    let self = this;
+    return this.store.findRecord('provider', 'admin').then(function(model) {
+      return model;
+    }).catch(function(reason) {
+      console.debug(reason);
 
-        self.get('flashMessages').warning('Fabrica is currently unavailable due to a DataCite API problem. We apologize for the inconvenience and are working hard to restore the service. Please check back later or contact DataCite Support if you have a question.');
-        self.transitionTo('index');
-      });
-    }
+      self.get('flashMessages').warning('Fabrica is currently unavailable due to a DataCite API problem. We apologize for the inconvenience and are working hard to restore the service. Please check back later or contact DataCite Support if you have a question.');
+    });
   },
 
   afterModel() {
-    if (this.can.cannot('read index') && this.currentUser) {
-      let home = this.currentUser.get('home');
-      if (home && home.id) {
-        this.transitionTo(home.route, home.id);
-      } else if (home) {
-        this.transitionTo(home.route);
-      } else {
-        this.transitionTo('index');
-      }
+    if (this.can.cannot('read index')) {
+      this.transitionTo('index');
     }
-  },
-
-  actions: {
-    queryParamsDidChange() {
-      this.refresh();
-    },
   },
 });
