--- app/routes/settings/index.js
@@ -1,23 +1,0 @@
-import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
-
-export default Route.extend({
-  can: service(),
-
-  model() {
-    let self = this;
-    return this.store.findRecord('provider', 'admin').then(function(model) {
-      return model;
-    }).catch(function(reason) {
-      console.debug(reason);
-
-      self.get('flashMessages').warning('Fabrica is currently unavailable due to a DataCite API problem. We apologize for the inconvenience and are working hard to restore the service. Please check back later or contact DataCite Support if you have a question.');
-    });
-  },
-
-  afterModel() {
-    if (this.can.cannot('read index')) {
-      this.transitionTo('index');
-    }
-  },
-});
