+++ app/templates/components/index-header.hbs
@@ -11,9 +11,6 @@
       <nav.link-to @route="index" class="nav-link">Info</nav.link-to>
     </nav.item>
     <nav.item>
-      <nav.link-to @route="settings" class="nav-link">Settings</nav.link-to>
-    </nav.item>
-    <nav.item>
       <nav.link-to @route="providers" class="nav-link">Members</nav.link-to>
     </nav.item>
     <nav.item>
@@ -29,6 +26,9 @@
     </nav.item>
     <nav.item>
       <nav.link-to @route="dois" class="nav-link">DOIs</nav.link-to>
+    </nav.item>
+    <nav.item>
+      <nav.link-to @route="dashboard" class="nav-link">Dashboard</nav.link-to>
     </nav.item>
   </BsNav>
 </div>
